+++ src/Sentry/SentryOptions.cs
@@ -423,9 +423,9 @@ public class SentryOptions
         return string.Equals(requestBaseUrl, _sentryBaseUrl.Value, StringComparison.OrdinalIgnoreCase);
     }
 
-    private Func<SentryEvent, Hint, SentryEvent?>? _beforeSend;
+    private Func<SentryEvent, SentryHint, SentryEvent?>? _beforeSend;
 
-    internal Func<SentryEvent, Hint, SentryEvent?>? BeforeSendInternal => _beforeSend;
+    internal Func<SentryEvent, SentryHint, SentryEvent?>? BeforeSendInternal => _beforeSend;
 
     /// <summary>
     /// Configures a callback function to be invoked before sending an event to Sentry
@@ -435,7 +435,7 @@ public class SentryOptions
     /// application a chance to inspect and/or modify the event before it's sent. If the
     /// event should not be sent at all, return null from the callback.
     /// </remarks>
-    public void SetBeforeSend(Func<SentryEvent, Hint, SentryEvent?> beforeSend)
+    public void SetBeforeSend(Func<SentryEvent, SentryHint, SentryEvent?> beforeSend)
     {
         _beforeSend = beforeSend;
     }
@@ -453,15 +453,15 @@ public class SentryOptions
         _beforeSend = (@event, _) => beforeSend(@event);
     }
 
-    private Func<SentryTransaction, Hint, SentryTransaction?>? _beforeSendTransaction;
+    private Func<SentryTransaction, SentryHint, SentryTransaction?>? _beforeSendTransaction;
 
-    internal Func<SentryTransaction, Hint, SentryTransaction?>? BeforeSendTransactionInternal => _beforeSendTransaction;
+    internal Func<SentryTransaction, SentryHint, SentryTransaction?>? BeforeSendTransactionInternal => _beforeSendTransaction;
 
     /// <summary>
     /// Configures a callback to invoke before sending a transaction to Sentry
     /// </summary>
     /// <param name="beforeSendTransaction">The callback</param>
-    public void SetBeforeSendTransaction(Func<SentryTransaction, Hint, SentryTransaction?> beforeSendTransaction)
+    public void SetBeforeSendTransaction(Func<SentryTransaction, SentryHint, SentryTransaction?> beforeSendTransaction)
     {
         _beforeSendTransaction = beforeSendTransaction;
     }
@@ -475,9 +475,9 @@ public class SentryOptions
         _beforeSendTransaction = (transaction, _) => beforeSendTransaction(transaction);
     }
 
-    private Func<Breadcrumb, Hint, Breadcrumb?>? _beforeBreadcrumb;
+    private Func<Breadcrumb, SentryHint, Breadcrumb?>? _beforeBreadcrumb;
 
-    internal Func<Breadcrumb, Hint, Breadcrumb?>? BeforeBreadcrumbInternal => _beforeBreadcrumb;
+    internal Func<Breadcrumb, SentryHint, Breadcrumb?>? BeforeBreadcrumbInternal => _beforeBreadcrumb;
 
     /// <summary>
     /// Sets a callback function to be invoked when a breadcrumb is about to be stored.
@@ -486,7 +486,7 @@ public class SentryOptions
     /// Gives a chance to inspect and modify the breadcrumb. If null is returned, the
     /// breadcrumb will be discarded. Otherwise the result of the callback will be stored.
     /// </remarks>
-    public void SetBeforeBreadcrumb(Func<Breadcrumb, Hint, Breadcrumb?> beforeBreadcrumb)
+    public void SetBeforeBreadcrumb(Func<Breadcrumb, SentryHint, Breadcrumb?> beforeBreadcrumb)
     {
         _beforeBreadcrumb = beforeBreadcrumb;
     }
@@ -1093,7 +1093,7 @@ public class SentryOptions
     /// </summary>
     /// <remarks>
     /// This option applies only to complex objects being added to Sentry events as contexts or extras, which do not
-    /// implement <see cref="IJsonSerializable"/>.
+    /// implement <see cref="ISentryJsonSerializable"/>.
     /// </remarks>
     public bool JsonPreserveReferences
     {