+++ test/Sentry.Tests/AttachmentHelper.cs
@@ -2,7 +2,7 @@ namespace Sentry.Tests
 {
     internal static class AttachmentHelper
     {
-        internal static Attachment FakeAttachment(string name = "test.txt")
+        internal static SentryAttachment FakeAttachment(string name = "test.txt")
         => new(
             AttachmentType.Default,
             new StreamAttachmentContent(new MemoryStream(new byte[] { 1 })),